emulator/requirements.txt
numpy
opencv-python
pyserial
pynput
